add build-only-r-s option to Makefile to not build builder image;  don't delete /home/helx if it doesn't exist; bump minor version

+++ root/helx/10-user.sh
@@ -94,7 +94,9 @@ fi
 
 if [[ "$USER" != "$DEFAULT_USER" ]]; then
   if [[ $DELETE_DEFAULT_USER_HOME_IF_UNUSED == "yes" ]]; then
-    echo "deleting /home/$DEFAULT_USER"
-    rm -rf /home/$DEFAULT_USER
+    if [ -d /home/$DEFAULT_USER ]; then
+      echo "deleting /home/$DEFAULT_USER"
+      rm -rf /home/$DEFAULT_USER
+    fi
   fi
 fi
